Rename into bundle.mjs
