+++ demo/client/dune
@@ -12,15 +12,13 @@
 
 (rule
  (alias client)
- ; Make the client rule depend on the generated file by melange.emit, so it needs to wait to be built and it's rebuilt when the file changes.
  (deps
   (alias_rec melange-app)
-  (:script bundle.mjs))
- (target bundled.js)
+  (:script build.mjs))
+ (target app.js)
  (action
   (progn
-   (run node %{script} app/demo/client/index.js %{target})
-   (run cp %{target} app/demo/client/)))
+   (run node %{script} app/demo/client/index.js %{target})))
  (enabled_if
   (= %{profile} "dev")))
 
@@ -29,7 +27,7 @@
  (deps
   (alias_rec melange-app)
   (:input create-from-fetch.jsx)
-  (:script bundle.mjs))
+  (:script build.mjs))
  (action
   (progn
    (run node %{script} %{input} "app/demo/client/")))
@@ -41,7 +39,7 @@
  (deps
   (alias_rec melange-app)
   (:input runtime-with-client.jsx)
-  (:script bundle.mjs))
+  (:script build.mjs))
  (action
   (progn
    (run node %{script} %{input} "app/demo/client/")))
